About Dataset
Since 2008, guests and hosts have used Airbnb to expand on traveling possibilities and present a more unique, personalized way of experiencing the world. Today, Airbnb became one of a kind service that is used and recognized by the whole world. Data analysis on millions of listings provided through Airbnb is a crucial factor for the company. These millions of listings generate a lot of data - data that can be analyzed and used for security, business decisions, understanding of customers' and providers' (hosts) behavior and performance on the platform, guiding marketing initiatives, implementation of innovative additional services and much more.  
This dataset has around 49,000 observations in it with 16 columns and it is a mix between categorical and numeric values.  
In This Dataset we will Explore and analyze the data to discover key understandings (not limited to these) such as :  
- What can we learn about different hosts and areas?
- What can we learn from predictions? (ex: locations, prices, reviews, etc) 
- Which hosts are the busiest and why? 
- Is there any noticeable difference of traffic among different areas and what could be the reason for it?